// src/api/generated/model/mgmtTargetTypeRequestBodyPost.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * hawkBit REST APIs
 * Eclipse hawkBit™ is a domain-independent back-end framework for rolling out software updates to constrained edge devices as well as more powerful controllers and gateways connected to IP based networking infrastructure.

 * OpenAPI spec version: v1
 */
import type { MgmtDistributionSetTypeAssignment } from './mgmtDistributionSetTypeAssignment';

export interface MgmtTargetTypeRequestBodyPost {
  /** The name of the entity */
  name: string;
  /** The description of the entity */
  description?: string;
  /** The colour of the entity */
  colour?: string;
  /** Target type key */
  key?: string;
  /** Array of distribution set types that are compatible to that target type */
  compatibledistributionsettypes?: MgmtDistributionSetTypeAssignment[];
}